More reliably notify of that a test couldn't be run
This might be super broken, but so is my brain right now. For mysterious and supposedly unrelated reasons, I'm incredibly tired! I'm trusting that this code is useful even if it needs improvements. A lot of it assumes that the `path` property of corresponding `TestFile`s will always match, but that might not happen if certain algorithm changes are made, the consequences of which would not be able to be anticipated.